Remove more duplicated code
