+++ erftools/preprocessing/__init__.py
@@ -7,8 +7,6 @@ from .era5.Download_ERA5Data import Download_ERA5_ForecastData
 from .era5.IO import write_binary_simple_ERF
 from .era5.IO import write_binary_vtk_structured_grid
 from .era5.IO import write_binary_vtk_cartesian_file
-from .era5.IO import find_latlon_indices
-from .era5.IO import find_erf_domain_extents
 from .era5.IO import write_binary_vtk_cartesian
 from .era5.Plot_1D import plot_1d
 from .era5.ReadERA5DataAndWriteERF_IC import ReadERA5_3DData
@@ -20,8 +18,6 @@ from .gfs.Download_GFSData import Download_GFS_ForecastData
 from .gfs.IO import write_binary_simple_ERF
 from .gfs.IO import write_binary_vtk_structured_grid
 from .gfs.IO import write_binary_vtk_cartesian_file
-from .gfs.IO import find_latlon_indices
-from .gfs.IO import find_erf_domain_extents
 from .gfs.IO import write_binary_vtk_cartesian
 from .gfs.Plot_1D import plot_1d
 from .gfs.ReadGFSDataAndWriteERF_IC import ReadGFS_3DData
